Add revision parameter to package and other related defines
'present' ensures some revision of the package is installed and does not upgrade
the package when a new version is available.

'<revision number>' ensures that if <revision number> of the package is
installed it is not upgraded. Due to limitations of the android tool, if
<revision number> is not installed, the latest revision will be installed.

Fixes #34.

'latest' ensures that the latest revision of the package is installed and
upgrades the package when a new version available.

Fixes #26.